requirements.txt
Django==4.2.7
asgiref==3.8.1
django-environ==0.11.2
gunicorn==20.1.0
pytz==2023.2
sqlparse==0.4.4
djangorestframework==3.15.1
django-cors-headers==3.12.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
boto3==1.26.23
django-storages==1.13.1
django-health-check==3.18.1
django-rest-auth==0.9.5
django-allauth==0.43.0
pillow==10.3.0
PyMySQL==1.1.0
tzdata

#mysqlclient==2.2.4


